@pcrespov refactor Meta entries
