+++ services/api-server/src/simcore_service_api_server/models/schemas/meta.py
@@ -1,14 +1,8 @@
-from pydantic import AnyHttpUrl, BaseModel, Field
+from models_library.api_schemas__common.meta import BaseMeta
+from pydantic import AnyHttpUrl
 
-from ..basic_types import VersionStr
 
-
-class Meta(BaseModel):
-    name: str
-    version: VersionStr
-    released: dict[str, VersionStr] | None = Field(
-        None, description="Maps every route's path tag with a released version"
-    )
+class Meta(BaseMeta):
     docs_url: AnyHttpUrl
     docs_dev_url: AnyHttpUrl
