+++ services/director-v2/src/simcore_service_director_v2/api/routes/meta.py
@@ -1,14 +1,15 @@
 from fastapi import APIRouter
-from models_library.api_schemas_directorv2.meta import Meta, VersionStr
+from models_library.api_schemas__common.meta import BaseMeta
+from models_library.basic_types import VersionStr
 
 from ...meta import API_VERSION, API_VTAG
 
 router = APIRouter()
 
 
-@router.get("", response_model=Meta)
-async def get_service_metadata() -> Meta:
-    return Meta(
+@router.get("", response_model=BaseMeta)
+async def get_service_metadata() -> BaseMeta:
+    return BaseMeta(
         name=__name__.split(".")[0],
         version=VersionStr(API_VERSION),
         released={API_VTAG: VersionStr(API_VERSION), "v0": VersionStr("0.1.0")},
